Improvements to the Guardfile

+++ Guardfile
@@ -4,6 +4,7 @@
 guard 'spork' do
   watch('config/application.rb')
   watch('config/environment.rb')
+  watch('config/cucumber.yml')
   watch(%r{^config/environments/.*\.rb$})
   watch(%r{^config/initializers/.*\.rb$})
   watch('Gemfile')
@@ -13,18 +14,15 @@ guard 'spork' do
   watch(%r{features/support/}) { :cucumber }
 end
 
-guard 'cucumber', :cli => "--drb --require features/" do
+guard 'cucumber', :cli => "--drb --require features/", :all_after_pass => true do
   watch(%r{^features/.+\.feature$})
   watch(%r{^features/support/.+$})                      { 'features' }
   watch(%r{^features/step_definitions/(.+)_steps\.rb$}) { |m| Dir[File.join("**/#{m[1]}.feature")][0] || 'features' }
+  watch('spec/factories.rb')                            { "cucumber" }
 
-  # Guard#start runs once right after startup
-  callback(:start_begin) { coverage }
-  callback(:start_end) { no_coverage }
-
-  # Guard#run_all runs
-  callback(:run_all_begin) { coverage }
-  callback(:run_all_end) { no_coverage }
+  # Guard#run_all runs when you hit 'Enter' in the terminal
+  callback(:run_all_begin) { coverage    }
+  callback(:run_all_end)   { no_coverage }
 end
 
 guard 'rspec', :version => 2, :all_after_pass => true do
@@ -46,13 +44,9 @@ guard 'rspec', :version => 2, :all_after_pass => true do
   # FactoryGirl factory definitions
   watch('spec/factories.rb')  { "spec" }
 
-  # Guard#start runs once right after startup
-  callback(:start_begin) { coverage }
-  callback(:start_end) { no_coverage }
-
-  # Guard#run_all runs
-  callback(:run_all_begin) { coverage }
-  callback(:run_all_end) { no_coverage }
+  # Guard#run_all runs when you hit 'Enter' in the terminal
+  callback(:run_all_begin) { coverage    }
+  callback(:run_all_end)   { no_coverage }
 end
 
 def coverage
@@ -60,5 +54,7 @@ def coverage
 end
 
 def no_coverage
-  `rm tmp/coverage.txt`
+  if File.exists?( 'tmp/coverage.txt' )
+    `rm tmp/coverage.txt`
+  end
 end
